refactor: 타입 분리 및 center 정렬 수정
overFlow일 땐 center 기능을 막는다.
