+++ src/lib/Tabs/hooks/useTabsScroll.tsx
@@ -1,29 +1,27 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { RefObject, useEffect, useLayoutEffect, useState } from 'react';
-import { TabsDirection } from '../Tabs';
+import { ScrollState, TabsDirection } from '../types';
 
 const useTabsScroll = (
   ref: RefObject<HTMLDivElement> | null,
-  position: TabsDirection
+  direction: TabsDirection
 ) => {
   const [isOverFlow, setIsOverFlow] = useState(false);
-  const [scrollState, setScrollState] = useState<
-    'start' | 'end' | 'both' | null
-  >(null);
+  const [scrollState, setScrollState] = useState<ScrollState>(null);
 
   const handleMoveScroll = (pos: 'start' | 'end') => {
     if (!ref || !ref.current) return;
 
     const { scrollLeft, clientWidth, scrollTop, clientHeight } = ref.current;
 
-    if (position === 'horizontal')
+    if (direction === 'horizontal')
       ref.current.scrollTo({
         left:
           pos === 'end' ? scrollLeft + clientWidth : scrollLeft - clientWidth,
         behavior: 'smooth',
       });
 
-    if (position === 'vertical')
+    if (direction === 'vertical')
       ref.current.scrollTo({
         top:
           pos === 'end' ? scrollTop + clientHeight : scrollTop - clientHeight,
@@ -47,13 +45,13 @@ const useTabsScroll = (
     else setScrollState('start');
 
     if (
-      position === 'horizontal' &&
+      direction === 'horizontal' &&
       Math.abs(scrollWidth - scrollLeft - clientWidth) < 5
     )
       setScrollState('end');
 
     if (
-      position === 'vertical' &&
+      direction === 'vertical' &&
       Math.abs(scrollHeight - scrollTop - clientHeight) < 5
     )
       setScrollState('end');
