+++ src/lib/Tabs/types/index.tsx
@@ -1,0 +1,13 @@
+import { ReactNode } from 'react';
+
+export type Label = string | number;
+
+export type PanelList = {
+  label: Label;
+  disabled: boolean;
+  contents: ReactNode;
+}[];
+
+export type TabsDirection = 'horizontal' | 'vertical';
+
+export type ScrollState = 'start' | 'end' | 'both' | null;
